Mark win-64 scipy packages as broken.
There is (as far as I know) nothing wrong with these packages in an
isolated sense. However when conda is configured according to the
conda-forge documentation: strict channel priority and higher priority
to conda-forge they lock users into the version of scipy provided by
conda-forge (1 for each python version).

As I understand it building scipy on Windows is a cumbersome and manual
process hence only 1 version of scipy has been built for each python
version. Thus users of conda-forge on Windows are locked into scipy 1.3.1
(1.3.2 for python 3.8), whereas it is my understanding that if these
packages are marked as broken conda will resolve to scipy from pkgs/main
which provides both newer and older versions of scipy.