Mark as broken libprotobuf package v21.x
Five package versions (21.1 to 21.5) have been released of the
libprotobuf package with incorrect version numbers.  The protobuf
repository switched to release versioning that omits the major
version number, allowing (for some reason) the python version of
the package to have a different major version than the c/c++
packages.

The bot takes the release number as the suggested package version
number and maintainers of libprotobuf didn't realize until
yesterday that this was not correct.

Corrected releases of the 5 packages will be made as soon as
possible.