Early CouchDB build may contain wrong openssl
The build of CouchDB bundles OpenSSL, so it could happen that at compile time OpenSSL 3 was used, but a runtime dependency forces OpenSSL 1.1.1
The latest build has the OpenSSL migration and is built for both, but earlier builds may fail if anything depends on OpenSSL 1.1.1
Windows is not affected because it ships binaries. OSX might be affected but I have no way to test this.