Mark `patchelf` version 0.18.0 broken (pt. 2)
This version of `patchelf` regularly corrupts libraries it is patching.
Upstream is aware of the issue, but it is not yet fixed. Also we
encountered this before and marked packages `broken`. However some new
ones were published recently.  So this marks the new ones broken to
avoid further issues due to this `patchelf` version.