Mark some builds of cppyy-cling as broken on macOS
cppyy-cling had insufficient testing and it turns out that these builds
are not functional at all, they just segfault or crash in other ways.
This is due to a mismatch in libcxx versions used. However, no matter
which libcxx version we install, these do not become functional, so this
cannot be patched with a metadata patch.