Archive qiskit-terra
qiskit-terra was the old name for qiskit before it reorganized. In version 0.x, `qiskit` was an empty package depending on `qiskit-terra`. With 1.0, `qiskit` became the main package and `qiskit-terra` was discontinued. Now qiskit 2.x is out and 1.x is out of support (except security fixes) and qiskit 0.x and qiskit-terra are unsupported. It is safe to archive qiskit-terra.